Move upload progress to center panel instead of modal dialog
